feat(cmdodcadd): Allow cmddocadd to also define an autoload path.
When cmddocadd defines a sub-command and autoloading is enabled, if
there is any associated path with that sub-command, the command will be
loaded from that path rather than one generated from the command name.

While this requires implementors to know before-hand the sub-commands
they are defining, it allows them to add documentation of that command's
existance to make help calls more useful.